move XHRpromise to own module
