+++ src/User/index.js
@@ -4,6 +4,8 @@
  * @module src/User
  */
 
+import XHRpromise from 'src/XHRpromise';
+
 /**
  * API endpoints.
  *
@@ -33,47 +35,6 @@ const UIPaths = {
     login: '/login/'
 };
 Object.freeze(UIPaths);
-
-/**
- * Creates and sends an XMLHttpRequest.
- *
- * @param {string} method - The method for the request.
- * @param {string} url - The URL for the request.
- * @param {Object} [opts] - Options for the request.
- * @param {string} [opts.body] - The request body.
- * @param {string} [opts.contentType] - The content type for the body.
- * @param {number?} [opts.successStatus=null] - The expected response status for
- * successful requests. If `null`, the promise resolves for any status.
- * @returns {Promise} Resolves with the completed request, or rejects with an
- * error.
- */
-function XHRpromise(method, url, opts = {}) {
-    const { body, contentType, successStatus = null } = opts;
-
-    return new Promise((resolve, reject) => {
-        const xhr = new XMLHttpRequest();
-        xhr.open(method, url);
-        xhr.withCredentials = true;
-
-        contentType && xhr.setRequestHeader('Content-Type', contentType);
-        xhr.onreadystatechange = () => {
-            if (xhr.readyState !== xhr.DONE) {
-                return;
-            }
-
-            if (successStatus !== null && xhr.status !== successStatus) {
-                return reject(new Error(
-                    `'${method} ${url}' failed: `
-                    + `expected status ${successStatus}, got ${xhr.status}`
-                ));
-            }
-
-            return resolve(xhr);
-        };
-
-        xhr.send(body);
-    });
-}
 
 /**
  * Represents user state.
